refactor:  add useSearchParams hook instead
