+++ src/app/cmyk/page.tsx
@@ -1,18 +1,20 @@
 import CMYKEditions from '@/components/CMYKEditions';
-import CMYKEditionsSkeleton from '@/components/CMYKEditions/Skeleton';
+import CMYKEditionsSkeleton from '@/components/CMYKEditions/CMYKEditionsSkeleton';
+import { getAllCMYKProjects } from '@/lib/api.server';
 import { getPageMetadata } from '@/lib/seo';
 import type { AppPage } from '@/lib/types';
 import { Suspense } from 'react';
 
 export const generateMetadata = () => getPageMetadata('CMYK');
 
-export const dynamic = 'auto';
+const CMYKPage: AppPage = async () => {
+  const projects = await getAllCMYKProjects({
+    next: { revalidate: 60 },
+  });
 
-const CMYKPage: AppPage = async ({ searchParams }) => {
   return (
     <Suspense fallback={<CMYKEditionsSkeleton />}>
-      {/* @ts-expect-error Server Component */}
-      <CMYKEditions edition={searchParams.edition} />
+      <CMYKEditions projects={projects} />
     </Suspense>
   );
 };
