+++ src/components/CMYKEditions/cmykVersions.ts
@@ -1,0 +1,6 @@
+export const cmykVersions = [
+  { version: 'cmyk-1', name: 'CMYK 1', edition: 1 },
+  { version: 'cmyk-2', name: 'CMYK 2', edition: 2 },
+  { version: 'cmyk-3', name: 'CMYK 3', edition: 3 },
+  { version: 'cmyk-4', name: 'CMYK 4', edition: 4 },
+] as const;
